fix: devide test code for fast 2ndary build
