+++ test/test_cxx17_string_view.cpp
@@ -1,4 +1,10 @@
-﻿#include "../include/string_split.hpp"
+﻿/*=============================================================================
+  Copyright (C) 2016 yumetodo
+
+  Distributed under the Boost Software License, Version 1.0.
+  (See http://www.boost.org/LICENSE_1_0.txt)
+=============================================================================*/
+#include "../include/string_split.hpp"
 #ifdef STRING_SPLIT_HAS_CXX17_STRING_VIEW
 #include "../iutest/include/iutest.hpp"
 #include "constant.hpp"
